third_party/WebKit/LayoutTests/media/autoplay-muted.html

<!DOCTYPE html>
<title>Test for autoplay of muted video</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<script>
    test(function() {
        assert_true(!!window.internals
            && !!window.internals.settings
            && !!window.internals.runtimeFlags
            && !!window.eventSender,
            "This test only works when run as a layout test!");
    }, "Prerequisites to running the rest of the tests");

    window.internals.settings.setMediaPlaybackRequiresUserGesture(true);
    window.internals.runtimeFlags.autoplayMutedVideosEnabled = true;
    testRunner.setAutoplayAllowed(true);

    function createMutedVideoElement() {
        var e = document.createElement('video');
        e.src = findMediaFile('video', 'content/test');
        e.muted = true;
        return e;
    }

    async_test(function(t) {
        var e = createMutedVideoElement();
        e.autoplay = true;

        var expectedEvents = [ 'canplay', 'play', 'playing'];
        var eventWatcher = new EventWatcher(t, e, expectedEvents);
        eventWatcher.wait_for(expectedEvents).then(
            t.step_func_done(function() {
                assert_false(e.paused);
            }));
    }, "Test that a muted video with an autoplay attribute autoplays.");

    promise_test(function() {
        return createMutedVideoElement().play();
    }, "Test that play() on a muted video succeeds without gesture.");

    promise_test(function(t) {
        var e = createMutedVideoElement();
        return e.play().then(t.step_func_done(function() {
            e.muted = false;
            assert_true(e.paused, "The video should be paused.");
        }));
    }, "Test that unmuting an autoplayed video without gesture pauses.");

    async_test(function(t) {
        var e = createMutedVideoElement();

        e.play().then(t.step_func(function() {
            eventSender.mouseDown();
            eventSender.mouseUp();
        }));

        document.onclick = t.step_func_done(function() {
            e.muted = false;
            assert_false(e.paused, "The video should not be paused.");
        });
    }, "Test that unmuting autoplayed video with gesture doesn't pause it.");

    promise_test(function(t) {
        testRunner.setAutoplayAllowed(false);
        return promise_rejects(
            t,
            new DOMException(
                'play() can only be initiated by a user gesture.',
                'NotAllowedError'),
            createMutedVideoElement().play());
    }, "Test that muted videos don't autoplay when the setting is disabled");
</script>
